[product] 레디스가 다운됐을때를 대비한 DB 이중화 추가

+++ product/src/main/java/me/sjlee/product/domain/repository/OptionPurchaseManageRepository.java
@@ -6,5 +6,5 @@ public interface OptionPurchaseManageRepository {
     boolean increasePurchaseCount(SalesOption salesOption, int purchaseCount);
     boolean decreasePurchaseCount(SalesOption salesOption, int purchaseCount);
     void initPurchaseCount(SalesOption salesOption);
-    int getCurrentPurchaseCount(SalesOption salesOption);
+    long getCurrentPurchaseCount(SalesOption salesOption);
 }
